Explicitly cast to the narrowing type - Address MSVC warnings about narrowing conversions - Every place capacity and CalculateCapacityForSize are used they are used as map_index_t, so modify that from size_t

+++ src/google/protobuf/micro_string.h
@@ -75,18 +75,18 @@ class PROTOBUF_EXPORT MicroString {
 
     absl::string_view view() const { return {payload, size}; }
     char* owned_head() {
-      ABSL_DCHECK_GE(capacity, kOwned);
+      ABSL_DCHECK_GE(capacity, static_cast<uint32_t>(kOwned));
       return reinterpret_cast<char*>(this + 1);
     }
 
     void SetExternalBuffer(absl::string_view buffer) {
       payload = const_cast<char*>(buffer.data());
-      size = buffer.size();
+      size = static_cast<uint32_t>(buffer.size());
     }
 
     void SetInitialSize(size_t size) {
       PoisonMemoryRegion(owned_head() + size, capacity - size);
-      this->size = size;
+      this->size = static_cast<uint32_t>(size);
     }
 
     void Unpoison() { UnpoisonMemoryRegion(owned_head(), capacity); }
@@ -94,7 +94,7 @@ class PROTOBUF_EXPORT MicroString {
     void ChangeSize(size_t new_size) {
       PoisonMemoryRegion(owned_head() + new_size, capacity - new_size);
       UnpoisonMemoryRegion(owned_head(), new_size);
-      size = new_size;
+      size = static_cast<uint32_t>(new_size);
     }
   };
